refactor FileUploadToTestCase
