+++ iaso/test.py
@@ -14,6 +14,7 @@ from django.contrib.auth import get_user_model
 from django.contrib.auth.models import AnonymousUser, Permission
 from django.contrib.contenttypes.models import ContentType
 from django.core.files import File
+from django.core.files.storage import default_storage
 from django.http import HttpResponse, StreamingHttpResponse
 from django.test import TestCase as BaseTestCase
 from django.urls import clear_url_caches
@@ -369,6 +370,28 @@ class APITestCase(BaseAPITestCase, IasoTestCaseMixin):
                 self.assertTrue(any(isinstance(data[field], t) for t in expected_type))
             else:
                 self.assertIsInstance(data[field], expected_type)
+
+
+class FileUploadToTestCase(TestCase, IasoTestCaseMixin):
+    """
+    Common setup for testing file upload_to functions that rely on account and user information.
+    On every test, the default_storage is cleared to avoid name conflicts.
+    """
+
+    def setUp(self):
+        # Preparing test data
+        account_1_name = "test account 1"
+        self.account_1, self.data_source_1, self.version_1, self.project_1 = (
+            self.create_account_datasource_version_project("source 1", account_1_name, "project 1")
+        )
+        account_2_name = "***///"
+        self.account_2, self.data_source_2, self.version_2, self.project_2 = (
+            self.create_account_datasource_version_project("source 2", account_2_name, "project 2")
+        )
+
+        # Removing all InMemoryFileNodes inside the storage to avoid name conflicts - some can be kept by previous test classes
+        default_storage._root._children.clear()  # see InMemoryFileStorage in django/core/files/storage/memory.py
+        super().setUp()
 
 
 class MockClamavScanResults:
