+++ iaso/tests/models/test_import_gpkg.py
@@ -1,28 +1,12 @@
-from django.core.files.storage import default_storage
 from django.core.files.uploadedfile import UploadedFile
 
 from iaso.models import ImportGPKG
-from iaso.test import IasoTestCaseMixin, TestCase
+from iaso.test import FileUploadToTestCase
 
 
-class ImportGpkgTestCase(TestCase, IasoTestCaseMixin):
+class ImportGpkgTestCase(FileUploadToTestCase):
     FILE_NAME = "minimal.gpkg"
     FILE_PATH = f"iaso/tests/fixtures/gpkg/{FILE_NAME}"
-
-    def setUp(self):
-        # Preparing test data
-        account_1_name = "test account 1"
-        self.account_1, self.data_source_1, self.version_1, self.project_1 = (
-            self.create_account_datasource_version_project("source 1", account_1_name, "project 1")
-        )
-        account_2_name = "***///"
-        self.account_2, self.data_source_2, self.version_2, self.project_2 = (
-            self.create_account_datasource_version_project("source 2", account_2_name, "project 2")
-        )
-
-        # Removing all InMemoryFileNodes inside the storage to avoid name conflicts - some can be kept by previous test classes
-        default_storage._root._children.clear()  # see InMemoryFileStorage in django/core/files/storage/memory.py
-        super().setUp()
 
     def test_upload_to_happy_path(self):
         # Upload with a user that belongs to a (correctly named) account
